Extract logic of payments plugin detection
